Refactor NodeDescriptionKeys management
This commit changes how NodeDescriptionKeys are managed. Actually there is no more a NodeDescriptionKeys protocol, since everything can be used as a key (this is because there is no way to enforce that a generic parameter is an enum).

The reasons of this refactor are mainly two
- It was very very cumbersome and error prone to make a complex enum RawRepresentable
- The amount of CPU required to transform a string into an enum and vice versa was high, and the developers were in charge of make it faster, which was not acceptable (and probably not possibile).

By using the string interpolation (e.g., “\(key)”) we speed up of an order of magnitude (according to some tests) the computation and we also remove the burden for the developers to make the enums RawRepresentable