Execute side effects after the update state
As discussed internally it seems that it is better to run the side effects after the updateState. By moving the side effects after the updateState we are able to have the updatedState when running custom logic in the sideEffects.

This commit alone is not enough to solve the issues we are facing since we have basically lost the previous state. Following commits will solve also this problem